+++ core/compiler/optimization.cpp
@@ -643,7 +643,7 @@ IntermediateBlock* ItermediateOptimizer::InlineMethod(IntermediateBlock* inputs)
       IntermediateMethod* mthd_called = program->GetClass(instr->GetOperand())->GetMethod(instr->GetOperand2());
       // checked called method to determine if it can be inlined
       if(CanInlineMethod(mthd_called, inlined_mthds, lbl_jmp_offsets)) {
-        int local_instr_offset = current_method->GetEntries()->GetParameters().size() + 1; // GetLastLocalOffset(current_method, outputs) + 2;
+        int local_instr_offset = (int)current_method->GetEntries()->GetParameters().size() + 1; // GetLastLocalOffset(current_method, outputs) + 2;
         if(current_method->HasAndOr() || mthd_called->HasAndOr()) {
           local_instr_offset++;
         }
